14 api implementation to query generic attributes

+++ InterfaceFactory/CommHandler4Modbus/src/test/java/communicator/modbus/impl/IOPdemoBOX.java
@@ -172,46 +172,46 @@ public class IOPdemoBOX {
 		
 		static void tstIOPMeter()
 		{
-			float fVal1 = (float) 0.0, fVal2 = (float) 0.0, fVal3 = (float) 0.0, fVal4 = (float) 0.0;
+			double fVal1 = (float) 0.0, fVal2 = (float) 0.0, fVal3 = (float) 0.0, fVal4 = (float) 0.0;
 			String  sVal1 = "0.0", sVal2 = "0.0", sVal3 = "0.0", sVal4 ="0.0";
 			
 				try {
 	 				mbRTU.setUnitIdentifier(rtuAddr);	
 				    LOG.info("\n@:Testing IOPMeterData: ");
 	  				Thread.sleep(25);
-	  				fVal1 = devIOPMeter.getValByGDPType("VoltageAC", "VoltageL1").getFloat32();
+	  				fVal1 = devIOPMeter.getValByGDPType("VoltageAC", "VoltageL1").getFloat64();
 	  				Thread.sleep(10);            
-	  				fVal2 = devIOPMeter.getValByGDPType("VoltageAC", "VoltageL2").getFloat32(); 
+	  				fVal2 = devIOPMeter.getValByGDPType("VoltageAC", "VoltageL2").getFloat64();
 	  				Thread.sleep(10);
-	  				fVal3 = devIOPMeter.getValByGDPType("VoltageAC", "VoltageL3").getFloat32();
+	  				fVal3 = devIOPMeter.getValByGDPType("VoltageAC", "VoltageL3").getFloat64();
 	  				Thread.sleep(10);
-	  				fVal4 = devIOPMeter.getValByGDPType("Frequency", "Frequency").getFloat32();
+	  				fVal4 = devIOPMeter.getValByGDPType("Frequency", "Frequency").getFloat64();
 	  				LOG.info("  VoltageAC L1,2,3/Frequency [V,Hz]: " + fVal1 + ",  " + fVal2 + ",  "
 	  						+ fVal3 + ",  " + fVal4 + "  ");
 	  				Thread.sleep(10);
-	  				fVal1 = devIOPMeter.getValByGDPType("VoltageAC", "VoltageACL1-L2").getFloat32();
+	  				fVal1 = devIOPMeter.getValByGDPType("VoltageAC", "VoltageACL1-L2").getFloat64();
 	  				Thread.sleep(10);
-	  				fVal2 = devIOPMeter.getValByGDPType("VoltageAC", "VoltageACL1-L3").getFloat32();
+	  				fVal2 = devIOPMeter.getValByGDPType("VoltageAC", "VoltageACL1-L3").getFloat64();
 	  				Thread.sleep(10);
-	  				fVal3 = devIOPMeter.getValByGDPType("VoltageAC", "VoltageACL2-L3").getFloat32();
+	  				fVal3 = devIOPMeter.getValByGDPType("VoltageAC", "VoltageACL2-L3").getFloat64();
 	  				LOG.info("  VoltageAC L12/13/23 [V]:           " + fVal1 + ",  " + fVal2 + ",  "
 	  						+ fVal3 + "  ");
 	  				Thread.sleep(10);
-	  				fVal1 = devIOPMeter.getValByGDPType("CurrentAC", "CurrentACL1").getFloat32();
+	  				fVal1 = devIOPMeter.getValByGDPType("CurrentAC", "CurrentACL1").getFloat64();
 	  				Thread.sleep(10);
-	  				fVal2 = devIOPMeter.getValByGDPType("CurrentAC", "CurrentACL2").getFloat32();
+	  				fVal2 = devIOPMeter.getValByGDPType("CurrentAC", "CurrentACL2").getFloat64();
 	  				Thread.sleep(10);
-	  				fVal3 = devIOPMeter.getValByGDPType("CurrentAC", "CurrentACL3").getFloat32();
+	  				fVal3 = devIOPMeter.getValByGDPType("CurrentAC", "CurrentACL3").getFloat64();
 	  				LOG.info("  CurrentAC L1/2/3 [A]:              " + fVal1 + ",  " + fVal2 + ",  "
 	  						+ fVal3 + "  ");
 	  				Thread.sleep(10);
-	  				fVal1 = devIOPMeter.getValByGDPType("PowerFactor", "PowerFactor").getFloat32();
+	  				fVal1 = devIOPMeter.getValByGDPType("PowerFactor", "PowerFactor").getFloat64();
 	  				Thread.sleep(10);
-	  				fVal2 = devIOPMeter.getValByGDPType("PowerFactor", "PowerFactorL1").getFloat32();
+	  				fVal2 = devIOPMeter.getValByGDPType("PowerFactor", "PowerFactorL1").getFloat64();
 	  				Thread.sleep(10);
-	  				fVal3 = devIOPMeter.getValByGDPType("PowerFactor", "PowerFactorL2").getFloat32();
+	  				fVal3 = devIOPMeter.getValByGDPType("PowerFactor", "PowerFactorL2").getFloat64();
 	  				Thread.sleep(10);
-	  				fVal4 = devIOPMeter.getValByGDPType("PowerFactor", "PowerFactorL3").getFloat32();
+	  				fVal4 = devIOPMeter.getValByGDPType("PowerFactor", "PowerFactorL3").getFloat64();
 	  				LOG.info("  Powerfactor tot/L1/L2/L3:          " + fVal1 + ",  " + fVal2 + ",  "
 	  						+ fVal3 + ",  " + fVal4 + "  ");
 	  				Thread.sleep(10); 
@@ -318,7 +318,7 @@ public class IOPdemoBOX {
 			
 			static void tstOMCCIWallbox()
 			{
-				float fVal1 = (float) 0.0, fVal2 = (float) 0.0, fVal3 = (float) 0.0;
+				double fVal1 = (float) 0.0, fVal2 = (float) 0.0, fVal3 = (float) 0.0;
 				String  sVal1 = "0.0", sVal2 = "0.0";
 				SGrEVStateType sgrEVState = null;
 				SGrOCPPStateType sgrOCPPState = null;
@@ -334,11 +334,11 @@ public class IOPdemoBOX {
 							 devOMCCIWallbox.setVal("Curtailment", "HemsCurrentLimit", String.valueOf(CurtailCurrent));
 							 LOG.info("  Setting HemsCurrentLimit to :     " + CurtailCurrent + "  ");
 						 }
-						 fVal1 = devOMCCIWallbox.getValByGDPType("CurrentAC", "CurrentACL1").getFloat32();
+						 fVal1 = devOMCCIWallbox.getValByGDPType("CurrentAC", "CurrentACL1").getFloat64();
 						 Thread.sleep(200);
-						 fVal2 = devOMCCIWallbox.getValByGDPType("CurrentAC", "CurrentACL2").getFloat32();
+						 fVal2 = devOMCCIWallbox.getValByGDPType("CurrentAC", "CurrentACL2").getFloat64();
 						 Thread.sleep(200);
-						 fVal3 = devOMCCIWallbox.getValByGDPType("CurrentAC", "CurrentACL3").getFloat32();
+						 fVal3 = devOMCCIWallbox.getValByGDPType("CurrentAC", "CurrentACL3").getFloat64();
 						 Thread.sleep(200);
 						 oEnumList = devOMCCIWallbox.getValByGDPType("EVSEState", "EV-StatusCode").getEnum();
 						 Thread.sleep(200);
@@ -351,19 +351,19 @@ public class IOPdemoBOX {
 						 LOG.info("  OCPP-StatusCode:                  " + sgrOCPPState + "  ");
 						 LOG.info("  CurrentAC[A]                      I[L1]= " + fVal1 + ",  I[L2] = "  + fVal2 + ",  I[L3] = "  + fVal3 + "  ");		 
 
-						 fVal1 = devOMCCIWallbox.getValByGDPType("ActivePowerAC", "ActivePowerACL1").getFloat32();
+						 fVal1 = devOMCCIWallbox.getValByGDPType("ActivePowerAC", "ActivePowerACL1").getFloat64();
 						 Thread.sleep(200);
-						 fVal2 = devOMCCIWallbox.getValByGDPType("ActivePowerAC", "ActivePowerACL2").getFloat32();
+						 fVal2 = devOMCCIWallbox.getValByGDPType("ActivePowerAC", "ActivePowerACL2").getFloat64();
 						 Thread.sleep(200);
-						 fVal3 = devOMCCIWallbox.getValByGDPType("ActivePowerAC", "ActivePowerACL3").getFloat32();
+						 fVal3 = devOMCCIWallbox.getValByGDPType("ActivePowerAC", "ActivePowerACL3").getFloat64();
 						 Thread.sleep(200);
 						 LOG.info("  PowerAC[kW]:                      P[1L]= " + fVal1 + ",  P[L2] = "  + fVal2 + ",  P[L3] = "  + fVal3 + "  ");	
 							 
-						 fVal1 = devOMCCIWallbox.getValByGDPType("ActiveEnergyAC", "ActiveEnergyACL1").getFloat32();
+						 fVal1 = devOMCCIWallbox.getValByGDPType("ActiveEnergyAC", "ActiveEnergyACL1").getFloat64();
 						 Thread.sleep(200);
-						 fVal2 = devOMCCIWallbox.getValByGDPType("ActiveEEnergyAC", "ActiveEnergyACL2").getFloat32();
+						 fVal2 = devOMCCIWallbox.getValByGDPType("ActiveEEnergyAC", "ActiveEnergyACL2").getFloat64();
 						 Thread.sleep(200);
-						 fVal3 = devOMCCIWallbox.getValByGDPType("ActiveEEnergyAC", "ActiveEnergyACL3").getFloat32();
+						 fVal3 = devOMCCIWallbox.getValByGDPType("ActiveEEnergyAC", "ActiveEnergyACL3").getFloat64();
 						 Thread.sleep(200);
 						 LOG.info("  EnergyAC[kWh] L1/L2/L3:           W[1] = " + fVal1 + "  W[2] = "  + fVal2 + "  W[3] = "  + fVal3 + "  ");	
 							
@@ -373,13 +373,13 @@ public class IOPdemoBOX {
 						 Thread.sleep(200);
 						 LOG.info("  EVState  support (ISO/IEC 15118): " + sVal1 + ",    EVCCID = " + sVal2 + "  ");
 						 
-						 fVal1 = devOMCCIWallbox.getValByGDPType("Curtailment", "SafeCurrent").getFloat32();
+						 fVal1 = devOMCCIWallbox.getValByGDPType("Curtailment", "SafeCurrent").getFloat64();
 						 Thread.sleep(200);
-						 fVal2 = devOMCCIWallbox.getValByGDPType("Curtailment", "HemsCurrentLimit").getFloat32();
+						 fVal2 = devOMCCIWallbox.getValByGDPType("Curtailment", "HemsCurrentLimit").getFloat64();
 						 Thread.sleep(200);
-						 fVal3 = devOMCCIWallbox.getValByGDPType("Curtailment", "HWCurrentLimit").getFloat32();
+						 fVal3 = devOMCCIWallbox.getValByGDPType("Curtailment", "HWCurrentLimit").getFloat64();
 						 Thread.sleep(200);
-						 iVal1 = devOMCCIWallbox.getValByGDPType("Curtailment", "maxReceiveTimeSec").getInt16U();
+						 iVal1 = (int)devOMCCIWallbox.getValByGDPType("Curtailment", "maxReceiveTimeSec").getInt16U();
 						 Thread.sleep(200);
 						 LOG.info("  Curtailment:                      SafeCurrent = " + fVal1 + "  HemsCurrentLimit = "  + fVal2 + "  HWCurrentLimit = "  + fVal3 +  "  maxReceiveTimeSec = "  + iVal1 +"  ");
